+++ raphael-solver/src/lib.rs
@@ -74,23 +74,19 @@ impl SolverSettings {
         self.simulator_settings.max_cp
     }
 
-    pub fn max_progress(&self) -> u32 {
-        #[allow(clippy::useless_conversion)]
-        u32::from(self.simulator_settings.max_progress)
+    pub fn max_progress(&self) -> u16 {
+        self.simulator_settings.max_progress
     }
 
-    pub fn max_quality(&self) -> u32 {
-        #[allow(clippy::useless_conversion)]
-        u32::from(self.simulator_settings.max_quality)
+    pub fn max_quality(&self) -> u16 {
+        self.simulator_settings.max_quality
     }
 
-    pub fn base_progress(&self) -> u32 {
-        #[allow(clippy::useless_conversion)]
-        u32::from(self.simulator_settings.base_progress)
+    pub fn base_progress(&self) -> u16 {
+        self.simulator_settings.base_progress
     }
 
-    pub fn base_quality(&self) -> u32 {
-        #[allow(clippy::useless_conversion)]
-        u32::from(self.simulator_settings.base_quality)
+    pub fn base_quality(&self) -> u16 {
+        self.simulator_settings.base_quality
     }
 }
